refactor(adapter): name responses tool adapter export
